refactor(core): move related code into feature
- `KeyframeSymbol`
- `reservedKeyframes`
