+++ packages/core/src/features/css-keyframes.ts
@@ -1,0 +1,40 @@
+import type { Imported } from './st-import';
+
+export interface KeyframesSymbol {
+    _kind: 'keyframes';
+    alias: string;
+    name: string;
+    import?: Imported;
+    global?: boolean;
+}
+
+export const reservedKeyFrames = [
+    'none',
+    'inherited',
+    'initial',
+    'unset',
+    /* single-timing-function */
+    'linear',
+    'ease',
+    'ease-in',
+    'ease-in-out',
+    'ease-out',
+    'step-start',
+    'step-end',
+    'start',
+    'end',
+    /* single-animation-iteration-count */
+    'infinite',
+    /* single-animation-direction */
+    'normal',
+    'reverse',
+    'alternate',
+    'alternate-reverse',
+    /* single-animation-fill-mode */
+    'forwards',
+    'backwards',
+    'both',
+    /* single-animation-play-state */
+    'running',
+    'paused',
+];
