+++ packages/core/src/native-reserved-lists.ts
@@ -100,37 +100,6 @@ export const nativePseudoElements = [
     'spelling-error',
 ];
 
-export const reservedKeyFrames = [
-    'none',
-    'inherited',
-    'initial',
-    'unset',
-    /* single-timing-function */
-    'linear',
-    'ease',
-    'ease-in',
-    'ease-in-out',
-    'ease-out',
-    'step-start',
-    'step-end',
-    'start',
-    'end',
-    /* single-animation-iteration-count */
-    'infinite',
-    /* single-animation-direction */
-    'normal',
-    'reverse',
-    'alternate',
-    'alternate-reverse',
-    /* single-animation-fill-mode */
-    'forwards',
-    'backwards',
-    'both',
-    /* single-animation-play-state */
-    'running',
-    'paused',
-];
-
 export const nativeFunctionsDic = {
     annotation: true,
     attr: true,
